fix(ffi): change RGMD radius from u32 to f32 and scale to match vertex space
Positions are scaled by /20.0 but radius was left as a raw u32 in engine
units. Now radius is f32 divided by the same ENGINE_UNIT_SCALE, so it
lives in the same coordinate space as vertex positions.

+++ src/ffi/types.rs
@@ -32,7 +32,7 @@ pub struct RgmdHeader {
     pub frame_count: i32,
     pub total_vertex_count: i32,
     pub total_index_count: i32,
-    pub radius: u32,
+    pub radius: f32,
 }
 // 28 bytes, align 4.
